+++ jdaviz/core/tests/test_tools.py
@@ -1,4 +1,5 @@
 import numpy as np
+from numpy.testing import assert_allclose
 from glue.core import Data
 
 
@@ -22,3 +23,27 @@ def test_boxzoom(cubeviz_helper, spectral_cube_wcs):
     t.interact.selected_y = [20, 60]
 
     assert t.get_x_axis_with_aspect_ratio() == (-5., 35.)
+
+
+def _get_lims(viewer):
+    return [viewer.state.x_min, viewer.state.x_max,
+            viewer.state.y_min, viewer.state.y_max]
+
+
+def test_rangezoom(specviz_helper, spectrum1d):
+    specviz_helper.load_data(spectrum1d, data_label='test')
+
+    sv = specviz_helper.app.get_viewer('spectrum-viewer')
+    assert_allclose(_get_lims(sv), [6000, 8000, 12.30618014327326, 16.542560043585965])
+
+    t = sv.toolbar.tools['jdaviz:xrangezoom']
+    t.activate()
+    t.interact.selected = [6500, 7000]
+    t.on_update_zoom()
+    assert_allclose(_get_lims(sv), [6500, 7000, 12.30618014327326, 16.542560043585965])
+
+    t = sv.toolbar.tools['jdaviz:yrangezoom']
+    t.activate()
+    t.interact.selected = [14, 15]
+    t.on_update_zoom()
+    assert_allclose(_get_lims(sv), [6500, 7000, 14, 15])